@@ -1,17 +1,19 @@
 package org.by1337.bparser.event;
 
+import net.minecraft.registry.entry.RegistryEntry;
+
 public class SoundEvent {
-    private net.minecraft.sound.SoundEvent sound;
+    private RegistryEntry<net.minecraft.sound.SoundEvent> sound;
     private float volume;
     private float pitch;
 
-    public SoundEvent(net.minecraft.sound.SoundEvent sound, float volume, float pitch) {
+    public SoundEvent(RegistryEntry<net.minecraft.sound.SoundEvent> sound, float volume, float pitch) {
         this.sound = sound;
         this.volume = volume;
         this.pitch = pitch;
     }
 
-    public net.minecraft.sound.SoundEvent getSound() {
+    public RegistryEntry<net.minecraft.sound.SoundEvent> getSound() {
         return sound;
     }
 
@@ -26,7 +28,7 @@ public class SoundEvent {
     @Override
     public String toString() {
         return "SoundEvent{" +
-                "sound=" + sound.getId().toString() +
+                "sound=" + sound.getKey().orElse(null) +
                 ", volume=" + volume +
                 ", pitch=" + pitch +
                 '}';
