@@ -1,30 +1,49 @@
 package org.by1337.bparser.listener;
 
+import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
-import net.fabricmc.fabric.api.client.command.v1.ClientCommandManager;
-import net.fabricmc.fabric.api.client.command.v1.FabricClientCommandSource;
-import net.minecraft.text.LiteralText;
+import net.fabricmc.fabric.api.client.command.v2.FabricClientCommandSource;
+import net.minecraft.text.MutableText;
 import net.minecraft.text.Text;
-import net.minecraft.text.TranslatableText;
+import net.minecraft.text.TextContent;
+import org.by1337.bparser.BParser;
 import org.by1337.bparser.cfg.Config;
 import org.by1337.bparser.event.GameMessageS2CPacketAccessor;
 import org.by1337.bparser.event.NetworkEvent;
 import org.by1337.bparser.text.RawMessageConvertor;
 import org.by1337.bparser.util.ChatUtil;
 
-import static net.fabricmc.fabric.api.client.command.v1.ClientCommandManager.argument;
+import static net.fabricmc.fabric.api.client.command.v2.ClientCommandManager.argument;
 
 public class ChatListener {
+    public ChatListener() {
+        NetworkEvent.CHAT_EVENT.register(packet -> {
+            if (!Config.INSTANCE.chat.clickToCopy)
+                return;
+            if (packet.content().getContent() instanceof TextContent) {
+                Text msg = packet.content();
+                MutableText literalText = Text.literal("").append(msg);
 
-    public void register() {
-        ClientCommandManager.DISPATCHER.register(LiteralArgumentBuilder.<FabricClientCommandSource>literal("//chat_click_to_copy")
+                String raw = Text.Serialization.toJsonString(packet.content());
+                String result = RawMessageConvertor.convert(raw);
+                ChatUtil.addCopyButton(literalText, " [copy]", result);
+                if (Config.INSTANCE.chat.asRaw) {
+                    ChatUtil.addCopyButton(literalText, "[raw]", raw);
+                }
+                ((GameMessageS2CPacketAccessor) (Object) packet).setMessage(literalText);
+            }
+        });
+    }
+
+    public void register(CommandDispatcher<FabricClientCommandSource> dispatcher) {
+        dispatcher.register(LiteralArgumentBuilder.<FabricClientCommandSource>literal("//chat_click_to_copy")
                 .executes(ctx -> {
                     Config.INSTANCE.chat.clickToCopy = !Config.INSTANCE.chat.clickToCopy;
                     if (Config.INSTANCE.chat.clickToCopy) {
-                        ctx.getSource().sendFeedback(new TranslatableText("lang.bparser.chat.on"));
+                        ctx.getSource().sendFeedback(Text.translatable("lang.bparser.chat.on"));
                     } else {
-                        ctx.getSource().sendFeedback(new TranslatableText("lang.bparser.chat.off"));
+                        ctx.getSource().sendFeedback(Text.translatable("lang.bparser.chat.off"));
                     }
                     Config.INSTANCE.save();
                     return 1;
@@ -33,9 +52,9 @@ public class ChatListener {
                         .executes(ctx -> {
                             Config.INSTANCE.chat.asRaw = !Config.INSTANCE.chat.asRaw;
                             if (Config.INSTANCE.chat.asRaw) {
-                                ctx.getSource().sendFeedback(new TranslatableText("lang.bparser.chat.raw.on"));
+                                ctx.getSource().sendFeedback(Text.translatable("lang.bparser.chat.raw.on"));
                             } else {
-                                ctx.getSource().sendFeedback(new TranslatableText("lang.bparser.chat.raw.off"));
+                                ctx.getSource().sendFeedback(Text.translatable("lang.bparser.chat.raw.off"));
                             }
                             Config.INSTANCE.save();
                             return 1;
@@ -45,9 +64,9 @@ public class ChatListener {
                         .executes(ctx -> {
                             Config.INSTANCE.chat.gradients = !Config.INSTANCE.chat.gradients;
                             if (Config.INSTANCE.chat.gradients) {
-                                ctx.getSource().sendFeedback(new TranslatableText("lang.bparser.chat.gradients.on"));
+                                ctx.getSource().sendFeedback(Text.translatable("lang.bparser.chat.gradients.on"));
                             } else {
-                                ctx.getSource().sendFeedback(new TranslatableText("lang.bparser.chat.gradients.off"));
+                                ctx.getSource().sendFeedback(Text.translatable("lang.bparser.chat.gradients.off"));
                             }
                             Config.INSTANCE.save();
                             return 1;
@@ -66,9 +85,9 @@ public class ChatListener {
                                     try {
                                         Config.INSTANCE.textType = Config.TextType.valueOf(input.toUpperCase());
                                         Config.INSTANCE.save();
-                                        ctx.getSource().sendFeedback(new TranslatableText("lang.bparser.chat.type", input));
+                                        ctx.getSource().sendFeedback(Text.translatable("lang.bparser.chat.type", input));
                                     } catch (IllegalArgumentException e) {
-                                        ctx.getSource().sendError(new TranslatableText("lang.bparser.chat.type.error", input));
+                                        ctx.getSource().sendError(Text.translatable("lang.bparser.chat.type.error", input));
                                         return 0;
                                     }
                                     return 1;
@@ -76,20 +95,5 @@ public class ChatListener {
                         )
                 )
         );
-
-        NetworkEvent.CHAT_EVENT.register(packet -> {
-            if (!Config.INSTANCE.chat.clickToCopy || !Thread.currentThread().getName().contains("Netty Client IO"))
-                return;
-            if (packet.getMessage() instanceof LiteralText) {
-                LiteralText literalText = (LiteralText) packet.getMessage();
-                String raw = Text.Serializer.toJson(packet.getMessage());
-                String result = RawMessageConvertor.convert(raw);
-                ChatUtil.addCopyButton(literalText, " [copy]", result);
-                if (Config.INSTANCE.chat.asRaw) {
-                    ChatUtil.addCopyButton(literalText, "[raw]", raw);
-                }
-                ((GameMessageS2CPacketAccessor) packet).setMessage(literalText);
-            }
-        });
     }
 }
