@@ -1,41 +1,28 @@
 package org.by1337.bparser.listener;
 
+import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
-import net.fabricmc.fabric.api.client.command.v1.ClientCommandManager;
-import net.fabricmc.fabric.api.client.command.v1.FabricClientCommandSource;
+import net.fabricmc.fabric.api.client.command.v2.ClientCommandManager;
+import net.fabricmc.fabric.api.client.command.v2.FabricClientCommandSource;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.entity.effect.StatusEffect;
-import net.minecraft.text.LiteralText;
+import net.minecraft.registry.Registries;
 import net.minecraft.text.MutableText;
-import net.minecraft.text.TranslatableText;
-import net.minecraft.util.registry.Registry;
+import net.minecraft.text.Text;
 import org.by1337.bparser.cfg.Config;
 import org.by1337.bparser.event.NetworkEvent;
 import org.by1337.bparser.util.ChatUtil;
 
 public class EffectListener {
-    public void register() {
-        ClientCommandManager.DISPATCHER.register(LiteralArgumentBuilder.<FabricClientCommandSource>literal("//effect_log")
-                .executes(ctx -> {
-                    Config.INSTANCE.effectLog = !Config.INSTANCE.effectLog;
-                    if (Config.INSTANCE.effectLog) {
-                        ctx.getSource().sendFeedback(new TranslatableText("lang.bparser.effect.on"));
-                    } else {
-                        ctx.getSource().sendFeedback(new TranslatableText("lang.bparser.effect.off"));
-                    }
-                    Config.INSTANCE.save();
-                    return 1;
-                })
-        );
-
+    public EffectListener() {
         NetworkEvent.MOB_EFFECT.register(packet -> {
-            if (!Config.INSTANCE.effectLog || !Thread.currentThread().getName().contains("Netty Client IO")) return;
+            if (!Config.INSTANCE.effectLog ) return;
             MinecraftClient mc = MinecraftClient.getInstance();
 
-            if (mc.player != null && packet.getEntityId() == mc.player.getEntityId()) {
-                StatusEffect effect = StatusEffect.byRawId(Byte.toUnsignedInt(packet.getEffectId()));
+            if (mc.player != null && packet.getEntityId() == mc.player.getId()) {
+                StatusEffect effect = packet.getEffectId();
                 StringBuilder sb = new StringBuilder();
-                String effectId = Registry.STATUS_EFFECT.getId(effect).getPath();
+                String effectId = Registries.STATUS_EFFECT.getId(effect).getPath();
                 String amplifier = Integer.toString(Byte.toUnsignedInt(packet.getAmplifier()));
                 String duration = Integer.toString(packet.getDuration());
                 sb.append("effect: ").append(effectId);
@@ -46,16 +33,30 @@ public class EffectListener {
                 sb.append(" showIcon: ").append(packet.shouldShowIcon());
 
 
-                MutableText text = new LiteralText("effect: ")
-                        .append(new LiteralText(effectId).styled(ChatUtil.copyText(effectId)))
-                        .append(", amplifier: ").append(new LiteralText(amplifier).styled(ChatUtil.copyText(amplifier)))
-                        .append(", duration: ").append(new LiteralText(duration).styled(ChatUtil.copyText(duration)));
+                MutableText text = Text.literal("effect: ")
+                        .append(Text.literal(effectId).styled(ChatUtil.copyText(effectId)))
+                        .append(", amplifier: ").append(Text.literal(amplifier).styled(ChatUtil.copyText(amplifier)))
+                        .append(", duration: ").append(Text.literal(duration).styled(ChatUtil.copyText(duration)));
 
                 ChatUtil.addCopyButton(text, sb.toString());
 
                 ChatUtil.show(text);
             }
         });
+    }
 
+    public void register(CommandDispatcher<FabricClientCommandSource> dispatcher) {
+        dispatcher.register(LiteralArgumentBuilder.<FabricClientCommandSource>literal("//effect_log")
+                .executes(ctx -> {
+                    Config.INSTANCE.effectLog = !Config.INSTANCE.effectLog;
+                    if (Config.INSTANCE.effectLog) {
+                        ctx.getSource().sendFeedback(Text.translatable("lang.bparser.effect.on"));
+                    } else {
+                        ctx.getSource().sendFeedback(Text.translatable("lang.bparser.effect.off"));
+                    }
+                    Config.INSTANCE.save();
+                    return 1;
+                })
+        );
     }
 }
