@@ -4,7 +4,6 @@ import net.minecraft.client.MinecraftClient;
 import net.minecraft.text.*;
 import net.minecraft.util.Formatting;
 
-import java.util.UUID;
 import java.util.function.UnaryOperator;
 
 public class ChatUtil {
@@ -12,9 +11,11 @@ public class ChatUtil {
     public static MutableText addCopyButton(MutableText text, String msg, String data) {
         return addCopyButton(text, msg, data, Formatting.GREEN);
     }
+
     public static MutableText addCopyButton(MutableText text, String msg, String data, Formatting color) {
-        return text.append(new LiteralText(msg).styled(copyText(data, color)));
+        return text.append(Text.literal(msg).styled(copyText(data, color)));
     }
+
     public static MutableText addCopyButton(MutableText text, String data) {
         return addCopyButton(text, " [copy]", data);
     }
@@ -22,13 +23,14 @@ public class ChatUtil {
     public static UnaryOperator<Style> copyText(String text) {
         return copyText(text, Formatting.GREEN);
     }
+
     public static UnaryOperator<Style> copyText(String text, Formatting color) {
         return s -> s.withClickEvent(new ClickEvent(ClickEvent.Action.COPY_TO_CLIPBOARD, text))
-                .withHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, new LiteralText("copy")))
+                .withHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, Text.literal("copy")))
                 .withColor(color);
     }
 
     public static void show(Text text) {
-        MinecraftClient.getInstance().inGameHud.addChatMessage(net.minecraft.network.MessageType.CHAT, text, UUID.randomUUID());
+        MinecraftClient.getInstance().inGameHud.getChatHud().addMessage(text);
     }
 }
